+++ src/Parser/Php/PhpMethodReceiverKind.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Lsp\Parser\Php;
+
+enum PhpMethodReceiverKind
+{
+    case Variable;
+    case This;
+    case ThisProperty;
+    case Other;
+}