+++ src/Parser/Php/PhpTypedVariableKind.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Lsp\Parser\Php;
+
+enum PhpTypedVariableKind
+{
+    case Parameter;
+    case Property;
+    case PromotedProperty;
+}